# DOCS/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Darpan What-If Simulator",
    "description": "Darpan Labs What-If Simulator API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/twin/chat": {
      "post": {
        "summary": "Twin Chat",
        "operationId": "twin_chat_twin_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwinChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/twin/decide": {
      "post": {
        "summary": "Twin Decide",
        "operationId": "twin_decide_twin_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwinDecideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwinDecideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/match": {
      "post": {
        "summary": "Match",
        "operationId": "match_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/pins": {
      "get": {
        "summary": "List Pins",
        "operationId": "list_pins_admin_pins_get",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Pins Admin Pins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/pin": {
      "post": {
        "summary": "Upsert Pin",
        "operationId": "upsert_pin_admin_pin_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Pin Admin Pin Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/pin/{name}": {
      "delete": {
        "summary": "Delete Pin",
        "operationId": "delete_pin_admin_pin__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Pin Admin Pin  Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/simulate": {
      "post": {
        "summary": "Simulate",
        "operationId": "simulate_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Candidate"
      },
      "Conditioning": {
        "properties": {
          "behavior_archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behavior Archetype"
          },
          "psychographic_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Psychographic Tags"
          },
          "demographic_profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demographic Profile"
          }
        },
        "type": "object",
        "title": "Conditioning"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MatchRequest": {
        "properties": {
          "cta_seq": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CTAStep"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta Seq"
          },
          "z_or_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Or User Id"
          },
          "conditioning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Conditioning"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MatchRequest"
      },
      "MatchResponse": {
        "properties": {
          "twin_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Twin Weights"
          },
          "primary_twin": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Twin"
          }
        },
        "type": "object",
        "required": [
          "twin_weights"
        ],
        "title": "MatchResponse"
      },
...
      "SimulateRequest": {
        "properties": {
          "cta_seq": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CTAStep"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta Seq"
          },
          "z_or_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Or User Id"
          },
          "task": {
            "type": "string",
            "enum": [
              "choose_product",
              "refine"
            ],
            "title": "Task"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioPatch"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "topk": {
            "type": "integer",
            "title": "Topk",
            "default": 5
          },
          "explain": {
            "type": "string",
            "enum": [
              "none",
              "blend",
              "per_twin"
            ],
            "title": "Explain",
            "default": "none"
          },
          "deterministic": {
            "type": "boolean",
            "title": "Deterministic",
            "default": true
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 17
          },
          "mixture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mixture"
              },
              {
                "type": "null"
              }
            ]
          },
          "conditioning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Conditioning"
              },
              {
                "type": "null"
              }
            ]
          },
          "use_pin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Pin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task",
          "scenarios"
        ],
        "title": "SimulateRequest"
      },
      "SimulateResponse": {
        "properties": {
          "by_scenario": {
            "items": {
              "$ref": "#/components/schemas/ScenarioResult"
            },
            "type": "array",
            "title": "By Scenario"
          },
          "twin_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Twin Weights"
          },
          "primary_twin": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Twin"
          },
          "sim_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sim Config"
          }
        },
        "type": "object",
        "required": [
          "by_scenario",
          "twin_weights",
          "sim_config"
        ],
        "title": "SimulateResponse"
      },
      "TwinChatRequest": {
        "properties": {
          "twin_id": {
            "type": "string",
            "title": "Twin Id"
          },
          "history": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "conditioning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Conditioning"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "twin_id",
          "prompt"
        ],
        "title": "TwinChatRequest"
      },
      "TwinChatResponse": {
        "properties": {
          "twin_id": {
            "type": "string",
            "title": "Twin Id"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          }
        },
        "type": "object",
        "required": [
          "twin_id",
          "reply"
        ],
        "title": "TwinChatResponse"
      },
      "TwinDecideRequest": {
        "properties": {
          "twin_id": {
            "type": "string",
            "title": "Twin Id"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 20
          },
          "conditioning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Conditioning"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "twin_id",
          "context",
          "candidates"
        ],
        "title": "TwinDecideRequest"
      },
      "TwinDecideResponse": {
        "properties": {
          "twin_id": {
            "type": "string",
            "title": "Twin Id"
          },
          "decision": {
            "$ref": "#/components/schemas/TwinDecision"
          }
        },
        "type": "object",
        "required": [
          "twin_id",
          "decision"
        ],
        "title": "TwinDecideResponse"
      },
...
    }
  }
}